Rename to `initSharedState`
